src/binance_service.py
import os
import logging
from dotenv import load_dotenv
from binance.client import Client
from binance.exceptions import BinanceAPIException

load_dotenv()

class BinanceService:
    def __init__(self):
        ...